Move LiveTrader to core
